chore: split BlueprintConfigManifestSettings across multiple files
